+++ Modules/NetworkingDsc/Examples/Resources/NetIPInterface/4-ConfigureNetIPInterface_DisableWeakHostSend.ps1
@@ -1,0 +1,18 @@
+<#
+    .EXAMPLE
+    Disable the weak host send IPv4 setting for the network adapter with alias 'Ethernet'.
+#>
+Configuration Example
+{
+    Import-DscResource -Module NetworkingDsc
+
+    Node localhost
+    {
+        NetIPInterface DisableWeakHostSend
+        {
+            InterfaceAlias = 'Ethernet'
+            AddressFamily  = 'IPv4'
+            WeakHostSend   = 'Disable'
+        }
+    }
+}